+++ sql/logs/update_2.2.3.sql
@@ -14,13 +14,4 @@
 -- You should have received a copy of the GNU General Public License
 -- along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-CREATE TABLE llx_dolibase_logs(
-	rowid INTEGER AUTO_INCREMENT PRIMARY KEY,
-	module_id INTEGER NOT NULL,
-	module_name VARCHAR(255) NOT NULL,
-	object_id INTEGER NOT NULL,
-	object_element VARCHAR(100) DEFAULT NULL,
-	action VARCHAR(255) NOT NULL,
-	datec DATETIME NOT NULL,
-	fk_user INTEGER NOT NULL
-);
+ALTER TABLE llx_dolibase_logs ADD object_element VARCHAR(100) DEFAULT NULL;